Make GTFS read and FileHandler public
